+++ src/Events/FolderDeleted.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace OiLab\OiLaravelAttachments\Events;
+
+use OiLab\OiLaravelAttachments\Models\Folder;
+
+/**
+ * Dispatched after a folder has been deleted (soft delete).
+ */
+class FolderDeleted
+{
+    public function __construct(
+        public readonly Folder $folder,
+    ) {}
+}